(PC-43001) feat(a11y): add version and debuggage to main section
